requirements.txt
pandas
numpy
scipy
streamlit
mysql-connector-python
plotly
